--- src/pages/configurePatch/configurePatchCore/ConfigureTasks/types.ts
@@ -1,9 +1,0 @@
-enum CheckboxState {
-  CHECKED = "CHECKED",
-  INDETERMINATE = "INDETERMINATE",
-  UNCHECKED = "UNCHECKED",
-  DISABLED_CHECKED = "DISABLED_CHECKED",
-  DISABLED_INDETERMINATE = "DISABLED_INDETERMINATE",
-}
-
-export { CheckboxState };
