+++ src/pages/configurePatch/configurePatchCore/ConfigureTasks/utils.test.ts
@@ -1,4 +1,4 @@
-import { CheckboxState } from "./types";
+import { CheckboxState } from "../types";
 import {
   deduplicateTasks,
   getSelectAllCheckboxState,
