+++ src/pages/configurePatch/configurePatchCore/index.tsx
@@ -26,16 +26,16 @@ import {
   ProjectBuildVariant,
 } from "gql/generated/types";
 import { SCHEDULE_PATCH } from "gql/mutations";
+import { ConfigureBuildVariants } from "./ConfigureBuildVariants";
+import ConfigureTasks from "./ConfigureTasks";
+import { ParametersContent } from "./ParametersContent";
 import {
   AliasState,
   ChildPatchAliased,
   PatchTriggerAlias,
   VariantTasksState,
-  useConfigurePatch,
-} from "hooks/useConfigurePatch";
-import { ConfigureBuildVariants } from "./ConfigureBuildVariants";
-import ConfigureTasks from "./ConfigureTasks";
-import { ParametersContent } from "./ParametersContent";
+} from "./types";
+import useConfigurePatch from "./useConfigurePatch";
 
 interface ConfigurePatchCoreProps {
   patch: ConfigurePatchQuery["patch"];
