+++ src/pages/configurePatch/configurePatchCore/types.ts
@@ -20,3 +20,11 @@ export type TasksState = {
 export type VariantTasksState = {
   [variant: string]: TasksState;
 };
+
+export enum CheckboxState {
+  CHECKED = "CHECKED",
+  INDETERMINATE = "INDETERMINATE",
+  UNCHECKED = "UNCHECKED",
+  DISABLED_CHECKED = "DISABLED_CHECKED",
+  DISABLED_INDETERMINATE = "DISABLED_INDETERMINATE",
+}
